potentially relevant for leap years: fixed unallocated 366th day in qout and qin potentially relevant for seasonality: faulty checking, could set some seasonal C-factors to 0
adjusted to meet standard of Intel Fortran Compiler 2010
- corrected inprecise INTENT statements
- implicite type casting in function arguments (soilwat)
- eliminated (some?) statements that relied on short-circuit-evaluation (which is not guaranteed in Fortran standard)

+++ trunk/Hillslope/sedi_yield.f90
@@ -79,7 +79,7 @@ SUBROUTINE sedi_yield(d, subbas_id, lu_id, tc_type_id, q_in, q_out, q_peak_in, v
     REAL, INTENT(IN):: q_out        !overland flow leaving the TC downslope [m**3]
     REAL, INTENT(IN):: q_peak_in    !peak overland flow rate  [m**3/s]
     REAL, INTENT(IN):: sed_in(1:n_sed_class)        !sediment entering the TC from uplope [tons/timestep] for each particle size class [currently not considered yet]
-    REAL, INTENT(IN):: timestep        !timestep which the given flows are related to [h]
+    INTEGER, INTENT(IN):: timestep        !timestep which the given flows are related to [h]
     REAL, INTENT(IN) :: tc_area        !area of TC [km**2]
     REAL, INTENT(IN) :: v_ov_in        !overland flow velocity [m/s]
     REAL, INTENT(OUT) :: sed_yield(1:n_sed_class)                !sediment yield [tons/timestep] (usually applied daily) for each particle size class
